Remove unused getInstalledApplications(Int) from PackageService
This change removes the unused `getInstalledApplications(flags: Int)` function from the `PackageService` interface and its implementation in `AndroidPackageService`. The `Long` version of the function is used by the application, so the `Int` version was dead code.

The `PackageServiceTest` was also updated to use the `Long` version of the function by casting `PackageManager.GET_META_DATA` to `Long`.

+++ app/src/test/kotlin/com/github/keeganwitt/applist/services/PackageServiceTest.kt
@@ -43,7 +43,7 @@ class PackageServiceTest {
 
         every { packageManager.getInstalledPackages(any<Int>()) } returns packages
 
-        val result = service.getInstalledApplications(PackageManager.GET_META_DATA)
+        val result = service.getInstalledApplications(PackageManager.GET_META_DATA.toLong())
 
         assertEquals(2, result.size)
         assertEquals("com.test.app1", result[0].packageName)
